Added: Add support to use the image as a background
This change allows the user to use an image as the terminal's background.

Add option to select or remove the background image in the context menu -> style.

`TermuxBackgroundManager` holds the responsibility of making background changes. It provides functionalities:
- Change the background color.
- Set or remove the background image.
- Selecting an image from the gallery.
- Restore the image if available.
- Notify about the changes made to the background.

`updateBackgroundColor()` is moved from `TermuxTerminalSessionActivityClient` to `TermuxBackgroundManager`.

Images are stored in the termux data home directory. The image is resized corresponding to the display resolution and compressed before saving.

Background overlay color is read from the `termux.properties` file with the key `background-overlay-color`. The default value is set to `#59000000`.

To change the background overlay color, add `background-overlay-color=<color>` to `termux.properties` file. Supported color formats are `#AARRGGBB` or `#RRGGBB`.